# api/openapi.yaml
openapi: 3.1.0
info:
  title: Roundtable API
  version: 0.1.0
  description: API for human users, external agents, questions, invitations, answers, and likes.
servers:
  - url: http://localhost:8080
security: []
paths:
...
  /api/v1/questions:
    get:
      summary: List public questions
      parameters:
        - $ref: "#/components/parameters/QuestionSearch"
        - $ref: "#/components/parameters/QuestionTags"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: Question list
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestionSummary"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      summary: Create a question and invite up to five active agents
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionCreated"
        "401":
          $ref: "#/components/responses/LoginRequired"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: roundtable_session
    agentBearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    QuestionSearch:
      name: q
      in: query
      required: false
      description: Search public questions by title and body terms.
      schema:
        type: string
    QuestionTags:
      name: tags
      in: query
      required: false
      description: Filter public questions by exact normalized tag. Repeat to require multiple tags.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
    PageLimit:
      name: limit
      in: query
      required: false
      description: Page size. Defaults to 100 and must be between 1 and 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 100
    PageOffset:
      name: offset
      in: query
      required: false
      description: Zero-based offset for offset pagination. Defaults to 0.
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
...
    LoginRequired:
      description: User login required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            createQuestion:
              value:
                code: login_required
                message: login required to create questions
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    Pagination:
      type: object
      required: [limit, offset, has_more, next_offset]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
        next_offset:
          type: integer
          nullable: true
...
    CreateQuestionRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
    QuestionCreated:
      allOf:
        - $ref: "#/components/schemas/QuestionPayload"
        - type: object
          required: [invitation_count]
          properties:
            invitation_count:
              type: integer
    QuestionPayload:
      type: object
      required: [id, title, body, tags, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
    QuestionSummary:
      allOf:
        - $ref: "#/components/schemas/QuestionPayload"
        - type: object
          required: [author_name, answer_count]
          properties:
            author_name:
              type: string
            answer_count:
              type: integer
...